Create Compare matrices
R's ability to deal with different data structures for comparisons does not stop at vectors. Matrices and relational operators also work together seamlessly!

Instead of in vectors (as in the previous exercise), the LinkedIn and Facebook data is now stored in a matrix called views. The first row contains the LinkedIn information; the second row the Facebook information. The original vectors facebook and linkedin are still available as well.

Instructions

Using the relational operators you've learned so far, try to discover the following:

When were the views exactly equal to 13? Use the views matrix to return a logical matrix.
For which days were the number of views less than or equal to 14? Again, have R return a logical matrix.