Create Take control!
In this exercise, you will combine everything that you've learned so far: relational operators, logical operators and control constructs. You'll need it all!

We've pre-defined two values for you: li and fb, denoting the number of profile views your LinkedIn and Facebook profile had on the last day of recordings. Go through the instructions to create R code that generates a 'social media score', sms, based on the values of li and fb.

Instructions

Finish the control-flow construct with the following behavior:

If both li and fb are 15 or higher, set sms equal to double the sum of li and fb.
If both li and fb are strictly below 10, set sms equal to half the sum of li and fb.
In all other cases, set sms equal to li + fb.
Finally, print the resulting sms variable.