Create Find the error
We went ahead and pre-loaded some code for you, but there's still an error. Can you trace it and fix it?

In times of despair, help with functions such as sum() and rev() are a single command away; simply execute the code ?sum and ?rev.

Instructions

Fix the error by including code on the last line. Remember: you want to call mean() only once!